fix: remove test cases, rename package name, improve some comment
