+++ java/Two Pointers/PairSumSortedBruteForce.java
@@ -1,0 +1,15 @@
+import java.util.Arrays;
+
+public class PairSumSortedBruteForce {
+    public  int[] pairSumSortedBruteForce(int[] nums, int target) {
+        int n = nums.length;
+        for (int i = 0; i < n; i++) {
+            for (int j = i + 1; j < n; j++) {
+                if (nums[i] + nums[j] == target) {
+                    return new int[] {i, j};
+                }
+            }
+        }
+        return new int[] {};
+    }
+}
